From Bill:
The attached patch rearranges things so the function
Fl::get_system_colors
parses the switches passed by -fg, -bg, or -bg2.  This allows you to
interpret these switches without using Fl_Window::show(argc,argv), which
is inconvenent for a program that displays many windows, none of which
are the "main" window.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@114 ea41ed52-d2ee-0310-a9c1-e6b18d33e121