Bill's changes for allowing checkboxes and radio items in menus to be toggled without removing the menu.  I modified it so that it also works when hitting return or space to activate the menu item.
A couple of things, though:

If the menu item's callback opens a window of any kind, things will get
weird.  This is probably OK because I doubt there is any code out there
that relies on a toggle button in a menu opening up another window when
selected.

For some reason, the redraw of the menu item causes the menuwindow and
menubar to also be redrawn, causes a bit of unsightly flicker on bigger
more complex menus.   Any idea how to fix this, Bill?

Carl


The only problem I can see is that the redra


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@636 ea41ed52-d2ee-0310-a9c1-e6b18d33e121