I changed a bit the way Fl_Group redraw :
The rectangular region around all widgets are clipped out, whatever the
boxtype. If the box is not rectangular, I draw a part of the group box just
under the widget, then I draw the widget. In this way, there are less
flickering even with non rectangular box ! (for example try test/boxtype)
I the widget is actually a group with a non rectangular box (typically a
group without box !), I go even further : I do not draw the parent's box behind
from the parent draw method, instead the child group will use its
parent box and draw it around its widgets.

The only remainding flickering are the outside labels ... We could also
improve this, using the same trick as for non rectangular widgets, but this
might be a bit too tricky ...


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@769 ea41ed52-d2ee-0310-a9c1-e6b18d33e121