These are some (hopefully minor) changes we keep needing to edit into each new version of fltk for use at Digital Domain.  These were found by running a diff between 1.0.6 and our inhouse version.
1. WinMain is moved to the fl_call_main.c file so that it can easily be
conditionally-compiled out (code copied from fltk 2.0, not tested!)

2. Fl_win32 ignores the value of window->xclass().  This seemed to cause
problems when windows had different xclass() in the same app and did not
have any useful effect on windows.  This value is changed to set the icon
of the app on X.

3. sgi pragmas added to x.H to turn off warnings from their broken xlib
header files.  Please tell me if this causes any sgi compilers to complain,
we should add more stuff to the #ifdef to stop that.

4. Fl_Gl_Window::ortho() changed so that text and images are not erased
if the origin is off the left/bottom of the window.

5. Small change to Fl_Input so that a click that gives it the focus also
selects all the text.  This is more windoze-like and it seems the users
really like it.

Note also that we had to set USE_XDBE to zero to get things to work on
the newest Irix machines.  I give up, Xdbe has been nothing but an impossible
and confusing mess and it is obvious that the X servers are not testing it
and thus it is suffering from software rot so it is unlikely to work anywhere
in the future.  Not sure if we want to switch it off in the official version,
though.


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.0@774 ea41ed52-d2ee-0310-a9c1-e6b18d33e121