Changed height of menubars back to 30.  Only Microsoft menubars are that tiny (and KDE when emulating Microsoft, and a smaller height makes it too small for those people who may wish to increase their fontsize to larger than 12.
To really fix this, we need to have a better widget layout system that
allows widgets to have a changeable minimum width and height and positions
widgets relative to each other.  Jim Hourihan's Fl_Form group widget has
the necessary functionalty, but I don't like the interface.  We ought to
look at something like this for a future version of FLTK.  We can actually
simulate this (I think) with judicious use of groups and resizables, but
we need some easy interface to it (based on the widget relationships) and
FLUID should be able to handle it automatically.  This type of system would
also allow widgets to resize based on their contents, something needed for
(decent looking) internationalization.

I'm not saying that absolute coordinate positioning should go away, but we
should have a modern system as the default.  Please comment on this!

Carl


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@883 ea41ed52-d2ee-0310-a9c1-e6b18d33e121