Modified fluid again to fix pointer incompatibily problems when doing "w = o" ...
This time should be the last fix ! "w = o" is done only for the last widget
that is at the top level in the current function. w is declared of the type
of that widget, and the return type of the function is set to it too.

There will still be a problem now if the user set the return time of the
function to something else than the last widget's type ... I guess in this case,
fluid should not write "return w" at the end, and should let the user make his
own "return whathever" since it must be something else that he wants to
return ?


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@901 ea41ed52-d2ee-0310-a9c1-e6b18d33e121