No changes to keyboard navigation, but I am still working on it.
Cleaned up Fl_Widget.H:

Got rid of the geti/seti functions and wrote them as real functions,
this made the code shorter in even the simplest programs (probably
because the savings of not passing an argument to the function more
than outweighed the additional code of one function per style entry).

Removed the need for Fl_Group to be a friend by making several
protected functions public.  This should also eliminate the need for a
lot of kludges with the Fl_X class used to get around protection of
the Fl_Window, but I have not fixed this yet.

Larry Charlton's fix so input/output can scroll due to position()
calls even if they don't have the focus.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@958 ea41ed52-d2ee-0310-a9c1-e6b18d33e121