One of my change broke a lot of thing, so I reversed the update_child method to its original behaviour, and instead in Fl_Button::handle, I do
  case FL_FOCUS:
  case FL_UNFOCUS:
    if (box()->fills_rectangle())
      damage(FL_DAMAGE_HIGHLIGHT);
    else
      parent()->damage(FL_DAMAGE_ALL, x(), y(), w(), h());

In this way, the NO_BOX is corretly unfocused (this will draw the background
behind to erase the dotted box)

I'm not sure this is very neat, but it works and does not seem to break anything
else.

I put Fl_Style_start.cxx in the makefile. Next stage is to make fluid use it
...


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@960 ea41ed52-d2ee-0310-a9c1-e6b18d33e121