Fixed Fl_Pixmap (constructor was written wrong)
Setting both FL_ALIGN_LEFT|FL_ALIGN_RIGHT on a label makes it cut
the left off if it is too big, but align to the left if it is
smaller.  This is very useful for displaying pathnames in buttons in
the file chooser!

Also does the same thing vertically with FL_ALIGN_TOP|FL_ALIGN_BOTTOM.

It now produces FL_KEYUP events, and the menubar uses them: tapping
Alt is a shortcut into/out of the menubar!  (To check for a shift
key being "typed" you must look for the 'up' event and also check
Fl::event_clicks() for non zero.  This is true if the same key is used
for two events in a row, and is thus only true if the key went down
and up without any intermediate keys)

Calls layout() on windows before they are shown (but after the theme
is loaded).  This should be useful for widgets that do automatic
layout, and gets rid of the need to call it yourself.

Nifty new resizable() demo program, as suggested by Curtis.  Shows a
realistic layout and some instructions.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@977 ea41ed52-d2ee-0310-a9c1-e6b18d33e121