Tiny change to Fl.H to get around bug in new gcc versions?
Change to code to try alternative cases of shortcuts suggested by
Yaroslav Volovich.  It should now work for foreign letters.  However
I was unable to get it to do anything no matter how I set the locale.
It may be better to fix this by hardcoding in the iso-8859-1 character
set and not use ctype at all, "locale" is a horrid travestry anyway and
has done more to *prevent* internationalization than to help it by making
it virtually impossible to test foreign components.


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.0@1001 ea41ed52-d2ee-0310-a9c1-e6b18d33e121