Here is an attempt to make the image+text labels draw as requested by James Dean Palmer.  This one does not use any flags and just changes the behavior when the alignment is zero.  Turning on only FL_ALIGN_INSIDE restores the older behavior (this was unintended but I left it, we could use this setting for other purposes).
I'm not sure if this is sufficient but try it.

The new code is all inside an if statement at the start of Fl_Widget::
draw_label() and can be easily altered to perhaps use more flags.

Also, a more permanant change is that alignment is not done by the
Fl_Image::draw() method anymore, it is done by Fl_Widget::draw_label(),
as James wrote his.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1046 ea41ed52-d2ee-0310-a9c1-e6b18d33e121