I made FL_NORMAL_SIZE into a public int variable.  You can now change it before any widgets are created, in order to change the default size of the fonts used.  The most useful thing that can be done is to set it to 12 so that the default font size matches Win32.
I also fixed the menubar drawing a bit so that a menubar with a FL_FLAT_BOX
draws correctly.

It should also be documented that "Fl::set_color(Fl_Color(15),0,0,128);"
can be used to get Win32-style highlighting of menus and text items.  This
is not portable to fltk 2.0.

With these changes it should be easier to get fltk to look just like Win32.


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.0@1062 ea41ed52-d2ee-0310-a9c1-e6b18d33e121