Fixed icon problems for KDE (also sync'd up 1.0 and 2.0 versions)
fl_file_chooser() now creates a Fl_FileChooser::CREATE chooser.
Maybe we should have another (defaulted) argument to fl_file_chooser()
to choose the mode?

Updated the MenuTitle widget in Fl_Menu.cxx to set a window box of
FL_DOWN_BOX (which generates an up box?!?) so that the *title* of
pop-up menus shows up with the default menu button box; this
probably needs to be updated further to set the window_box() when
displaying the menu, with the window_box() being inherited from
the menu widget? (Bill?)


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1074 ea41ed52-d2ee-0310-a9c1-e6b18d33e121