Some bugfixes to get Windows to be happier.
FLTK now just keeps track of one pen and one brush at a time.  Cleaned
up some pen and brush memory leaks.

fL_getDC() now never puts the handle for the root window in fl_gc.  This
is just to make sure the FLTK doesn't accidentally try to dray on the root
instead of one of its windows.

Fl_Fontsize::Fl_Fontsize() was not storing the encoding, so every time a
font was used the encoding would not match any in the linked list and a new
Fl_Fontsize object was created.  This was the huge resource leak that was
sending Windows to la-la land.  (This cured the redrawing slowdown I had
noticed before.  Now things are nice and fast.)

Added a static struct with a destructor to Fl_win32.cxx to clean up Windows
resources when the program exits.  Not everything is being cleaned up yet,
though.

Fixed resizing of Windows by calling window->layout() when WM_SIZE is
received.  I'm not sure if this is the proper thing to do, but it works.

I haven't looked for the region code memory leak yet.  Hopefully this was
cured by Bill's last commit?

There are still some minor glitches, but the Windows port is now in a
reasonably usable state ands works pretty well!

Carl


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1098 ea41ed52-d2ee-0310-a9c1-e6b18d33e121