window->hide() now unmaps the window, like people expect.  Use window->destroy() to actually get rid of the window.  Known incompatabilites:
window->shown() will remain true after hide() is called.  This will
break loops that wait for shown() to turn off.

window->visible() is true when the window is iconized, and the user
iconizing the window does not cause hide/show events to be sent to the
window.  The new window->iconic method can be used to test if a window
is currently iconic.

The default callback for a window calls destroy() rather than hide(),
this is necessary so that modal windows can be reused with different
parents (both X and Win32 do not allow.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1195 ea41ed52-d2ee-0310-a9c1-e6b18d33e121