Okay, a better fix for the multiple child windows not showing up because I found more programs with bugs:
The rules for handle() are now that returning non-zero means no other
widgets should get the event.  Widgets should return 0 for FL_SHOW or
FL_HIDE (most did this already), if they return 1 then the event is
not sent to other widgets.  Also all widgets should pass events they
don't understand to their base class with BaseClass::handle(event).

Fluid is fixed so the windows show (it was not passing FL_SHOW to the
base window class, this is a common problem in fltk 1.0 programs)

Cube demo program fixed for this, also the sliders go the right way
again.

Fixed Fl_XPM_Image to allow arbitrary sized input (it does not crash
on the large demo one we got).


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1218 ea41ed52-d2ee-0310-a9c1-e6b18d33e121