Buttons with box(FL_NO_BOX) did not draw.  Apparently they did in older versions of fltk, I restored this. (bug 108771)
Removed 8-bit colormap drawing code that was not doing anything in
fl_draw_image due to Mike's changes.  I also made fl_color(r,g,b)
actually allocate the requested color rather than the nearest fltk
color-cube color (this is only done for the first color that maps to a
given entry in the fltk color cube), the result is that pixmaps with a
small number of colors are drawn much more accurately. The resulting
code seems to produce better images and is a good deal smaller!

Fixed makeinclude.in so CFLAGS are used for c source code instead of
CXXFLAGS. (bug 108694)

Better fix for gif files suggested by pauly (bug 108770)

Performance of Fl_Gl_Window may be improved on some types of OpenGL
  implementations, in particular MESA or other software emulators, by
  setting the GL_SWAP_TYPE environment variable.  This variable
  declares what is in the back buffer after you do a swapbuffers.

  setenv GL_SWAP_TYPE COPY

    This indicates that the back buffer is copied to the front buffer,
    and still contains it's old data. This is true of many hardware
    implementations.  Setting this will speed up emulation of
    overlays, and widgets that can do partial update can take
    advantage of this as damage() will not be cleared to -1.

  setenv GL_SWAP_TYPE NODAMAGE

    This indicates that nothing changes the back buffer except drawing
    into it.  This is true of MESA and Win32 software emulation and
    perhaps some hardware emulation on systems with lots of memory.

  All other values for GL_SWAP_TYPE, and not setting the variable,
  cause fltk to assumme that the back buffer must be completely
  redrawn after a swap.

  This is easily tested by running the gl_overlay demo program and
  seeing if the display is correct when you drag another window over
  it or if you drag the window off the screen and back on. You have to
  exit and run the program again for it to see any changes to the
  environment variable.


git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.0@1246 ea41ed52-d2ee-0310-a9c1-e6b18d33e121