Made forms compatability work with new style stuff, though maybe I should not have renamed draw_label to draw_inside_label, which was the main reason things did not work.
Child windows are not destroyed but only hidden on FL_HIDE. This puts
things exactly the way they are in fltk 1.0 (which actually works
pretty well, so perhaps the attempt to not destroy the windows was a
mistake?)

ConfigureNotify & MapNotify are ignored for child windows.

Swapped allocation of black & white in the hope this will fix the
overlays on Solaris.

PS: Thanks Carl for figuring out what was wrong with the tooltips!


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1284 ea41ed52-d2ee-0310-a9c1-e6b18d33e121