Cleaned up the code in Fl_compose a bit so the table is readable.
Added some "const" to the datas for Fl_Shared_Image

Lots of work trying to get fluid to work with the new browser, and to
clean up the code some. This now works, but I am still not happy with
how the browser and menus keep track of the current item and am still
trying to think of a solution.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1298 ea41ed52-d2ee-0310-a9c1-e6b18d33e121