Changed Fl_Function_Type.cxx in Fluid so that correct code is generated for functions which do NOT return the outermost widget.  For example, if you start a new Fluid project, add a function, create a window in that function, change the return type to bool, add code to return true, and generate code, you wind up with:
bool make_window()
{
   bool *w;
   { Fl_Window *o = new Fl_Window(0, 0, 100, 100, "Test");
     w = o;   // not valid!
   };
   return true;
}

Which doesn't compile.  The problem is in assuming the return type of the
function is the same as the type of the outermost widget, which it isn't.
I just changed it so that if the return type of the function is NOT the same as
the subclassname() of the last_group for the function, it uses the
subclassname() as the variable type for w.

(Simple change, but I thought the long explanation might be important.)


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1371 ea41ed52-d2ee-0310-a9c1-e6b18d33e121