More changes for Win32 to fix build errors and minor problems.
Fixes to Visual Studio 6 project files.  Added images project to workspace
and split out the files from the main fltk dlls.  THIS STILL NEEDS WORK as
the dependencies between the main dll and the image dll appear to be
circular (the image dll needs stuff from the main dll, but the main dll
needs stuff from the image dll).  Also haven't added the forms project or
the GL project yet.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1376 ea41ed52-d2ee-0310-a9c1-e6b18d33e121