Tooltips and message windows no longer are reused.  This was causing their styles not to work as expected.
For this change it was best to get rid of

Fl_Widget *fl_message_icon()

(which I've never seen actually used anyway).  I replaced it with

void fl_message_icon(Fl_Widget *) which allows the use to create their
own widget for the icon if they choose.  This is much more flexible as
it allows complex widgets or widgets with callbacks.

The message boxes now size themselves automatically to the icon, so any
size widget can be used.

fl_choice(...) now lets you choose which button is the default by starting
its string with the character '*' .  (The '*' is not displayed.)

Carl


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1388 ea41ed52-d2ee-0310-a9c1-e6b18d33e121