Updated the Visual Studio projects to add two new targets for each one of them: Debug MinSize and Release MinSize.  Both of these new targets use the "Multithreaded DLL" code generation option (which will link the MS runtime as a DLL rather than statically).  The existing targets were not changed.
Corrected a couple of problems in Fluid (a pointer that was getting trashed
in the call that sets the image root directory, since it was a stack buffer,
and needed to be static, and a linkage problem with missing __argc and __argv).

Added _MSC_DLL preprocessor define in Fluid project and in fl_call_main.c,
since apparently the MS runtime DLL doesn't provide those two symbols.  (They
are provided as empty in fl_call_main.c if _MSC_DLL is defined and the
fl_call_main version of WinMain is used.  Probably should really parse out
the command line and pass those instead, but this at least lets it build.)
As far as I know, _MSC_DLL is invented.  If there's an actual define to tell
you you're building for the DLL runtime, I don't know what it is.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1400 ea41ed52-d2ee-0310-a9c1-e6b18d33e121