Added to the fl_file_chooser() function on the Win32 platform so you can use the Windows stock file requester if you like.  Default is to use the FLTK one (same as it was before), and you can change it at runtime.
Added a function on Win32 called use_windows_stock_file_requester(), with a
boolean (int) argument that defaults to true.  Call this before calling
fl_file_chooser() and you'll get the stock Windows file requester.  It's just
a flag, so it stays that way until you turn it off again (which you can do
any time you want).

I'm not sure how the other developers feel about this, but it's one of my
boss's pet peeves about FLTK, and it only took some 30 lines of code to add
the ability to get the Windows requester.  (I realize one of FLTK's goals is
to provide the same look and feel on all platforms, but at least in our case,
the Windows users won't care what it looks like on other platforms, just what
it looks like on Windows -- and they have expectations of what the file
requester is supposed to look like.)   If this was a no-no, just let me
know...


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1403 ea41ed52-d2ee-0310-a9c1-e6b18d33e121