Added progress bar to windows project (source file was already there, it just wasn't in the project) for VC++.
Changed the Fl_Input_Browser so that when you set a callback, the callback gets set for the input part *and* the menu part.  However, something's still not quite right about Fl_Input_Browser's callbacks -- if you pull down the list and choose something, the callback doesn't fire.  I haven't been able to pin this one down just yet.

Fixed a code generation problem with Fluid where a class with a static member fucntion would get written as "static bool static function()" or whatever return type it was.

Changed the Fluid image loading code so that it fixes a strcpy() on an overlapped memory region, and so that it doesn't try to make relative paths out of absolute paths on Win32.  (That is, if the path is already C:\something, it doesn't tyy to turn it into D:\myproject\mydir\c:\something).


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1467 ea41ed52-d2ee-0310-a9c1-e6b18d33e121