Fl_Input_Browser somehow got broken.  Fixed.
Moved further along in font abstraction.  Not quite done.  Currently Xft
plugin is not compiling.  Will fix.

Now when a menu window is dismissed by clicking outside of it, the button
press is sent to the window/widget that it would have been sent to if the
pointer had not been grabbed.  Currently, this is only implemented on X,
will get around to Windows (should be easier).

More stuff that I forgot.

Carl


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1479 ea41ed52-d2ee-0310-a9c1-e6b18d33e121