Some '@' signs were ending up in fltk-config. There still is @LD_PLUGIN_FLAGS@ in it, but I could not figure this one out.
FL_HIDE was not being passed down to children of groups. This meant
that Fl_Windows that were in an Fl_Tab did not disappear when the tab
was switched, amoung other problems.

Fl_Menu_ callbacks further simplified. This gets rid of the static
Fl_Menu::callback_menu pointer I put in the last version, and allows a
program to emulate fltk1.0 with the addition of a callback. When a
menu or browser item is selected, the callback for the menu/browser is
done, and item() of that menu/browser is set to the selected item.
There is also Fl_Menu_::default_callback which is installed by the
constructor, it calls the callback on the item(). Thus by default it
does it the way Robert and fluid wants, but by replacing the menu's
callback you can alter the behavior so that things other than the item
and it's data are passed to the item's callback. Some demo programs
fixed to work with this.

FL_DIAMOND_UP_BOX and FL_DIAMOND_DOWN_BOX were swapped.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1554 ea41ed52-d2ee-0310-a9c1-e6b18d33e121