Fl_Image::draw() has an added Fl_Flags argument to support image types that draw different depending on the pushed (FL_VALUE), active, or highlighted state of the widget.
Fl_Image::draw() is also simplified to take just an x,y coordinate for
the upper-left corner. If you want to draw a portion of an image you
should use fl_push_clip before calling it.

I took out the indirection pointers for Fl_Image as that will not
work, because it bypasses the virtual draw function on Fl_Image. We
either need to make the _draw() function able to print an offscreen
image (pixmap/bitmap) onto another device, or the rendering context
should set a flag that each Fl_Image subclass should check to see
whether it should draw directly or try to create the offscreen images.

The "image" demo program rewritten to be much nicer and it can be used
to test Fl_Bitmap, Fl_Pixmap, and Fl_RGB_Image. Other image types can
be added fairly easily if wanted. Deleted the old bitmap and pixmap
demos. Some small bugs fixed in the clipping of images in labels.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@1559 ea41ed52-d2ee-0310-a9c1-e6b18d33e121