fl_background() will produce usable gray shades when the color is white or black (makes the high-contrast Windows styles work).
Lots of fixes to the WIN32 drawing code to get rid of blinking and
make the windows update when the application is not active.

Greg's rotation slider added to the symbols demo.

Added Fl_Multi_Image and multiimage demo program.

Valuator reported step() as Infinity, which messed up fluid's saved
files. Added ability to set line & page size in fluid. Fixed snap-to
behavior of sliders.

X version will produce characters when Cryillic letters are typed,
though they are only correct if koi8-r encoding is used.

Restored size and color of text for Fl_Value_Slider (note that this
also controls the size of the numbers next to the tick marks).


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@2211 ea41ed52-d2ee-0310-a9c1-e6b18d33e121