Fl_Input DnD changed to match OS/X: there is a small timeout before it starts DnD, if the user moves the mouse quickly it instead starts a new selection.
DnD in the same Fl_Input moves the text rather than copying it. Use
any shift key to force it to copy.

DnD bug when a group responded to FL_DND_DRAG but cursor is pointing
at an inner widget that responds to FL_DRAG. It would send the
FL_DND_RELEASE to the inner widget instead of the group.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@2312 ea41ed52-d2ee-0310-a9c1-e6b18d33e121