'Anonymous' reported a problem when using Fl_Value_Input. It seemed, that contrary to the man pages, a fractional 'step' value above 1 would not create a floating point input.
In fact it does, but the output was formatted wrong so that the 
digits after the decimal point were never rendered.

I changed the Fl_Valuator::format(double v) function how I beleive
it should format the output correctly, but as already stated by
the original author in the source code, this is a hack that should
be fixed by providing a 'precission' setting for valuators.

Anyway. My fix makes sure that all digits right of the decimal 
point are always rendered, so that the step value will show full
precision. This gives a much better behavior in respect to steps values
like 2.5, 3.75, etc., but also leads to 8 digits after the decimal point
for step(1.0/3.0)... .

I suggest that we keep this change (hence the commit), risking that 
rendering of valuator text will change in a few cases (odd step() values).



git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.1@4091 ea41ed52-d2ee-0310-a9c1-e6b18d33e121