Added ARGB32 image data type, to match the word-oriented libraries that are used on Windows. Removed some other pixel types that are (I think) not used often.
Image objects and drawimage() scale! OS/X does it perfectly, Win32 is
ok but there is no filtering. X does not do it, but perhaps XRender will
solve this.

Image objects scale to fill the rectangle. This is most useful if you
use them as the box() of a widget. Still need to use the inset() amounts
to scale edges differently than the interior.

drawimage() scales according to the current transform. However no
rotations or reflections yet. Possibly this is impossible on Win32.


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@4408 ea41ed52-d2ee-0310-a9c1-e6b18d33e121