Much improved sensing if a machine will do Alpha blending. This may actually be the solution. If Alpha blending is not supported, we fall back to the old screen door transparency.
This does not take into account a multiple screen setup with different screen depths, but I assume that this is a rare enough constellation and would cause many other problems on top of this minor issue. As a fallback, we will *always* render the image, but worst case without any alpha channel.

git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-1.1@5437 ea41ed52-d2ee-0310-a9c1-e6b18d33e121