Further shortcut fixes:
Removed previous fix as it just moves the type problems up. Put in the
usual workaround of int(long(x)) to get around casting bug in 64-bit
VC++ compiler.

Shortcut case is preserved and used to decide how to display in the menus.
It is ignored when matching shortcuts. Removed old kludge to control the
capitalization in menus, fixed some of the programs that used lowercase.

Punctuation matches even if shift is held down, so you don't have to
specify shift for punctuation marks (you can if you want to).


git-svn-id: http://seriss.com/public/fltk/fltk/trunk@6518 ea41ed52-d2ee-0310-a9c1-e6b18d33e121