Get test1 folder (mostly) building again...
I've been messing about in the test1 folder to get the fltk-1 tests building
under fltk3.

To that end, I've messed about with the Makefile - I've been testing with
Msys/mingw, and the version I have does not have a "makedepend" utility, 
so I have tweaked things to auto-generate the dependencies "by other means".

This works, but is probably not a long-term solution. Fine for testing though.

Also some very trivial fixes to some problems I stumbled over.

With these "fixes" in place, I now build 50 executables in the test1 folder.
The "native" fltk3 build creates 70 executables in the test folder, so we
appear to be 20 executables short of a full house.

The failing issue largely seem to pertain to issues with Fl_Bitmap, though
a few other problems are encountered too.

Hope this doesn't actualy break anything else!



git-svn-id: http://seriss.com/public/fltk/fltk/branches/branch-3.0@9297 ea41ed52-d2ee-0310-a9c1-e6b18d33e121